rename optimizer rule
