+++ src/frontend/src/optimizer/rule/agg_project_merge_rule.rs
