+++ src/frontend/src/optimizer/rule/apply_agg_transpose_rule.rs
