+++ src/frontend/src/optimizer/rule/apply_join_transpose_rule.rs
