+++ src/frontend/src/optimizer/rule/apply_project_transpose_rule.rs
