+++ src/frontend/src/optimizer/rule/apply_scan_rule.rs
