+++ src/frontend/src/optimizer/rule/index_delta_join_rule.rs
