+++ src/frontend/src/optimizer/rule/max_one_row_eliminate_rule.rs
