+++ src/frontend/src/optimizer/rule/merge_multijoin_rule.rs
