+++ src/frontend/src/optimizer/rule/project_eliminate_rule.rs
