+++ src/frontend/src/optimizer/rule/union_to_distinct_rule.rs
